// src/web-ui/src/app/models/attendance-work-report.model.ts
export class AttendanceWorkReportModel {
  id?: number;
  userId: number;
  date: any;
  remoteWork: boolean;
}

export class DicAttendanceWorkType {
  id?: number;
  name: string;
}

export class AttendanceWorkDay {
  id?: number;
  attendanceWorkReport: AttendanceWorkReportModel;
  dicAttendanceWorkType: DicAttendanceWorkType;
  hours: number;
  comment: string;

  constructor(attendanceWorkReport?: AttendanceWorkReportModel) {
    this.attendanceWorkReport = attendanceWorkReport;
  }
}